requirements.txt
streamlit
google-generativeai
python-dotenv
Pillow
python-docx
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-auth
chromadb
langchain-community
langchain-google-genai
